Constructors for custom weapons
Added a constructor that allows you to name a new weapon of each of the types. Created a custom weapon constructor that also takes in a die of static type for full customization. See weapons class file for example on how to create a full custom weapon until I figure out how to fix it.